test: check for leaks when connection setup not completed
When we intercept a SYN we `connect()` to the remote and send back
a SYN ACK. If the client responds with ACK then the connection is marked
as completed (and could use keep-alives etc). If the client never
responds with an ACK then the connection has effectively leaked.

This test emulates such a bad client and fails if the connection is
not closed within 30s.

Signed-off-by: David Scott <dave.scott@docker.com>