Avoid leaking fd on ethernet socket close
Previously if the peer closes the ethernet socket (e.g. a hyperkit process
shuts down) then there would be a read error on the socket and we would log

> read EOF so closing connection

but we wouldn't close the connection (unlike in the "unexpected" error case).
This patch calls the idempotent function `disconnect` which wakes up callers
blocked in `after_disconnect` which includes the caller which will close
the underlying connection.

Signed-off-by: David Scott <dave.scott@docker.com>