Add support for AF_UNIX SOCK_DGRAM
Unfortunately these aren't supported by libuv directly, so we need to
implement our own separate system. It doesn't need to be particularly
scalable as the expected use-case is to send and receive ethernet frames.

The server will listen on a Unix domain socket and receive connected
SOCK_DGRAM sockets which will have ethernet frames on them.

Signed-off-by: David Scott <dave@recoil.org>